nativesdk-sdk-provides-dummy: allow pkgconfig to be installed into SDKs
This recipe is a dummy provider of pkgconfig, but we want our SDKs to
ship a pkgconfig binary if requested as otherwise that's a host tool
that would be required to use the SDK. On Linux that's easily solved,
but meta-mingw/meta-darwin SDKs can't really assume that pkgconfig will
be present on the host.

This behaviour dates back to 2014 when construction of SDKs with RPM
packages was added[1] and notably the same behaviour didn't exist for
ipkg-based SDKs.

[1] oe-core 417b27ce5c1 ("lib/oe/sdk.py: support RpmRootfs")

(From OE-Core rev: 4aaceb859812d3fd9fdb3ae878515d0a799f9aa8)

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>